setting up standalone zap to be included in iframe
