+++ src/components/content/zus-aggregated-profile/zus-aggregated-profile.tsx
@@ -75,7 +75,16 @@ const ZusAggregatedProfileComponent = ({
   medicationsOutsideProps,
   medicationsAllProps,
   timelineProps,
-  resources,
+  resources = [
+    "conditions-all",
+    "medications-all",
+    "diagnostic-reports",
+    "timeline",
+    "documents",
+    "allergies",
+    "immunizations",
+    "care-team",
+  ],
   hideTitle = false,
   title = "Outside Records",
   removeBranding = false,
